Try with a hostfile to work around openmpi / macOS issue:
While trying to create a regular expression of the node names
used in this application, the regex parser has detected the
presence of an illegal character in the following node name:

  node:  sat12-bq146_ccbd4e97-5ee0-4b30-b5f6-75640051be87-86A39A8B25A3

Node names must be composed of a combination of ascii letters,
digits, dots, and the hyphen ('-') character. See the following
for an explanation:

  https://en.wikipedia.org/wiki/Hostname

Please correct the error and try again.
--------------------------------------------------------------------------
--------------------------------------------------------------------------
An internal error has occurred in ORTE:

[[36102,0],0] FORCE-TERMINATE AT (null):1 - error base/plm_base_launch_support.c(555)

This is something that should be reported to the developers.